Do not apply calibration on audio routed to afe proxy
When compress offload playback is active, if the Visualizer effect
is enabled, decoded PCM audio is routed to AFE Proxy port and
read from it by the Visualizer wrapper library. When audio is routed
to proxy port, current output device specific calibration is also
being applied which is not desired.
Avoid this by sending default audio calibration i.e. no post-processing
to be applied the data.

Change-Id: Id576c4ed7bbb482683074e3e33aa5760b7597d37